requirements.txt
fastapi==0.115.12
pymongo==4.11.3
motor===3.7.0
uvicorn==0.34.0
python-dotenv
firebase-admin==6.4.0